+++ neo-cli/CLI/MainService.Mode.cs
@@ -16,34 +16,13 @@ using System.IO;
 using System.IO.Compression;
 using System.Linq;
 using System.Net;
+using System.Text;
 using System.Threading.Tasks;
+using Akka.Util.Internal;
 namespace Neo.CLI;
 
  partial class MainService
     {
-        /// <summary>
-        /// Process "install" command
-        /// </summary>
-        /// <param name="modeName">Mode name</param>
-        [ConsoleCommand("mode load", Category = "Mode Commands")]
-        private async Task OnLoadMode(string modeName)
-        {
-
-
-            // using (response)
-            // {
-            //     await using Stream stream = await response.Content.ReadAsStreamAsync();
-            //     using ZipArchive zip = new(stream, ZipArchiveMode.Read);
-            //     try
-            //     {
-            //         zip.ExtractToDirectory(".");
-            //     }
-            //     catch (IOException)
-            //     {
-            //     }
-            // }
-        }
-
         /// <summary>
         /// Process "mode list" command
         /// <param name="modeName">Mode name</param>
@@ -51,28 +30,13 @@ namespace Neo.CLI;
         [ConsoleCommand("mode list", Category = "Mode Commands")]
         private void OnListModes()
         {
-            // var plugin = Plugin.Plugins.FirstOrDefault(p => p.Name == pluginName);
-            // if (plugin is null)
-            // {
-            //     ConsoleHelper.Warning("Plugin not found");
-            //     return;
-            // }
-            // if (plugin is Logger)
-            // {
-            //     ConsoleHelper.Warning("You cannot uninstall a built-in plugin.");
-            //     return;
-            // }
-
-            // File.Delete(plugin.Path);
-            // File.Delete(plugin.ConfigFile);
-            // try
-            // {
-            //     Directory.Delete(Path.GetDirectoryName(plugin.ConfigFile), false);
-            // }
-            // catch (IOException)
-            // {
-            // }
-            // ConsoleHelper.Info("Uninstall successful, please restart neo-cli.");
+            try
+            {
+                Directory.GetDirectories("./Modes/").ForEach(p => ConsoleHelper.Info(p));
+            }
+            catch (IOException)
+            {
+            }
         }
 
         /// <summary>
@@ -82,19 +46,7 @@ namespace Neo.CLI;
         [ConsoleCommand("mode save", Category = "Mode Commands")]
         private void OnSaveMode(string modeName)
         {
-            // if (Plugin.Plugins.Count > 0)
-            // {
-            //     Console.WriteLine("Loaded plugins:");
-            //     foreach (Plugin plugin in Plugin.Plugins)
-            //     {
-            //         if (plugin is Logger) continue;
-            //         ConsoleHelper.Info($"\t{plugin.Name,-20}", plugin.Description);
-            //     }
-            // }
-            // else
-            // {
-            //     ConsoleHelper.Warning("No loaded plugins");
-            // }
+
         }
 
         /// <summary>
@@ -116,6 +68,62 @@ namespace Neo.CLI;
             // else
             // {
             //     ConsoleHelper.Warning("No loaded plugins");
+            // }
+        }
+
+
+
+        private void LoadConfig(string mode)
+        {
+            var dir = new DirectoryInfo(mode);
+            if (!dir.Exists)
+                throw new DirectoryNotFoundException($"Mode not found: {dir.FullName}");
+
+            // Cache directories before we start copying
+            var dirs = dir.GetDirectories();
+
+            // Get the config files of the node
+            foreach (var file in dir.GetFiles())
+            {
+                var targetFilePath = Path.Combine("./", file.Name);
+                file.CopyTo(targetFilePath, true);
+            }
+
+            // Copy the Plugin files
+            foreach (var plugin in dirs)
+            {
+                foreach (var file in plugin.GetFiles())
+                {
+                    var targetFilePath = Path.Combine($"Plugins/{plugin.Name}", file.Name);
+                    file.CopyTo(targetFilePath, true);
+                }
+            }
+        }
+
+        private void SaveMode(string mode)
+        {
+            string path = @"./Modes/Mode";
+
+            try
+            {
+                // if(File.Exists(path)) File.Delete(path);
+                using (var fs = File.Create(path))
+                {
+                    byte[] info = new UTF8Encoding(true).GetBytes(mode);
+                    fs.Write(info, 0, info.Length);
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.ToString());
+            }
+            // using (StreamReader sr = File.OpenText(path))
+            // {
+            //     string s = "";
+            //     while ((s = sr.ReadLine()) != null)
+            //     {
+            //         Console.WriteLine(s);
+            //     }
             // }
         }
     }
