+++ neo-cli/Modes/Mode
@@ -1,0 +1,1 @@
+mainnet
